Tidying

+++ archivelib-rs-tdd/src/compress/fn211.rs
@@ -1,8 +1,8 @@
 use crate::compress::{CompressU16ArrayAlias, CompressU8ArrayAlias, RCompressData, Result};
 
 use crate::support::ArrayAlias;
-use std::io::Read;
 use crate::support::BitwiseWrite;
+use std::io::Read;
 
 impl<R: Read, W: BitwiseWrite> RCompressData<R, W> {
   pub fn fn211(
@@ -11,12 +11,12 @@ impl<R: Read, W: BitwiseWrite> RCompressData<R, W> {
     dat_arr_cursor187: &mut CompressU16ArrayAlias<'_>,
     dat_arr_cursor178: &mut CompressU8ArrayAlias<'_>,
     dat_arr_cursor188: &mut CompressU16ArrayAlias<'_>,
-  ) -> Result<i32> {
+  ) -> Result<u32> {
     let dat_arr_cursor178_offset = dat_arr_cursor178.offset(self);
     let dat_arr_cursor188_offset = dat_arr_cursor188.offset(self);
 
     self.dat174 = var212 as i16;
-    let mut var292 = var212;
+    let mut var292 = var212 as u32;
     let mut var227 = 0;
     self.dat_arr177[1] = 0;
     for i in 0..(self.dat174 as usize) {
@@ -28,7 +28,7 @@ impl<R: Read, W: BitwiseWrite> RCompressData<R, W> {
     }
     if var227 < 2 {
       dat_arr_cursor188.set(self, self.dat_arr177[1] as usize, 0);
-      return Ok(self.dat_arr177[1] as i32);
+      return Ok(self.dat_arr177[1] as u32);
     } else {
       let mut run_start226 = (var227 / 2) as i16;
       while run_start226 >= 1 {
@@ -67,7 +67,7 @@ impl<R: Read, W: BitwiseWrite> RCompressData<R, W> {
         }
       }
       dat_arr_cursor188.set_offset(self, dat_arr_cursor188_offset);
-      self.fn228(var289, dat_arr_cursor178, dat_arr_cursor188);
+      self.fn228(var289 as i32, dat_arr_cursor178, dat_arr_cursor188);
       dat_arr_cursor188.set_offset(self, dat_arr_cursor188_offset);
       dat_arr_cursor178.set_offset(self, dat_arr_cursor178_offset);
       self.fn230(var212, dat_arr_cursor178, dat_arr_cursor188);
